Update fallacies.json
Most of the global logic community calls the fallacy "post hoc ergo propter hoc", shortened to either "post hoc fallacy" or "causalation". I know you're based in the Turkish-speaking community and all, but I never see "previous this" unless it's adjacent to its Turkish name and usually also the Latin name familiar to logicians around the world. And on the Anglophone-facing version of the site it is best to reflect a universal consensus on the name.